backend/requirements.txt
Django>=3.2
djangorestframework
django-cors-headers
gunicorn
psycopg2
